requirements.txt
streamlit==1.8.1
numpy==1.22.4
keras==2.8.0
Pillow==8.4.0
